UIOR-1068 update code structure
