+++ src/common/utils/index.js
@@ -1,3 +1,4 @@
+export * from './checkPOLinesAbandonedHoldings';
 export * from './checkRelatedHoldings';
 export * from './createClearFilterHandler';
 export * from './getAcqUnitsByIds';
@@ -7,6 +8,7 @@ export * from './getAddresses';
 export * from './getAddressOptions';
 export * from './getCommonErrorMessage';
 export { default as getCreateInventorySetting } from './getCreateInventorySetting';
+export * from './getHoldingPiecesAndItemsCount';
 export * from './getOrganizationsByIds';
 export * from './getLocations';
 export * from './getMaterialTypes';
